examples/wikipedia: always send atleast 1 message
The wikipedia driver was originally coded to produce messages with a
random number generator. The generator was for the range [0,99] which
meant there was a chance that no records are produced. This caused a
problem in Redpanda integration testing where success is determined by
parsing output.

This commit ensures that atleast 1 record is always sent to the example
topics.

Fixes: https://github.com/redpanda-data/redpanda/issues/2889